% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterRecommend.R
\name{addCluster}
\alias{addCluster}
\title{The addCluster function}
\usage{
addCluster(joinedData, userProfile, joinVariable = "user_id")
}
\arguments{
\item{joinedData}{dataframe of all reviews, which includes business and
other user information}

\item{userProfile}{data frame that contains a user_id and cluster assignment}

\item{joinVariable}{(optional) variable that the left_join will be based on}
}
\value{
dataframe that has review, business, user and cluster assignment
}
\description{
Adds (left_joins) the user_id's cluster assignment to the full review dataset
}
\examples{
data("join100Sample")
data("user100Sample")
clusters <- 4
userFeatures <- appendUserData(fulldata = join100Sample, userdata = user100Sample)
userWithCluster <- appendKmeansCluster(userFeatures, k = clusters)
busWithCluster <- addCluster(join100Sample,userWithCluster)
length(names(join100Sample))
length(names(busWithCluster))
rm("join100Sample","user100Sample","clusters","userFeatures","userWithCluster","busWithCluster")
}
\author{
Ken Calhoon
}
\keyword{data}

