% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansClustering.R
\name{appendKmeansCluster}
\alias{appendKmeansCluster}
\title{The appendKmeansCluster function}
\usage{
appendKmeansCluster(userdata, k = 4)
}
\arguments{
\item{userdata}{dataframe of user data}

\item{k}{(optional) number of clusters}
}
\value{
dataframe with cluster assignments appended
}
\description{
Appends to the input userdata file the cluster assignments
}
\examples{
data("join100Sample")
data("user100Sample")
userFeatures <- appendUserData(fulldata = join100Sample, userdata = user100Sample)
userCluster <- appendKmeansCluster(userFeatures, k = 4)[,c("user_id","user_name","km_cluster")]
head(userCluster)
table(userCluster$km_cluster)
rm("join100Sample","user100Sample","userFeatures","userCluster")
}
\author{
Ken Calhoon
}
\keyword{data}

