% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{appendUserData}
\alias{appendUserData}
\title{The appendUserData function}
\usage{
appendUserData(fulldata, userdata)
}
\arguments{
\item{fulldata}{dataframe with all reviews, business and basic user data}

\item{userdata}{dataframe with user profile data. New variables will be 
appended to this dataframe}
}
\value{
dataframe with new user features appended to the input "userdata"
}
\description{
Wrapper function that appends user features. Uses \code{\link{convertDate}},
 \code{\link{featActiveDays}}, \code{\link{featPredFemale}}, 
 and \code{\link{featSumBusDataByUser}}.
}
\examples{
data("join100Sample")
data("user100Sample")
dim(user100Sample)
names(user100Sample)
u <- appendUserData(fulldata = join100Sample, userdata = user100Sample)
dim(u)
names(u)
rm("u", "join100Sample", "user100Sample")
}
\author{
Ken Calhoon
}
\keyword{data}

