% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus_sample_list_data.R
\docType{data}
\name{busSampleListData}
\alias{busSampleListData}
\title{Sample business data from Yelp Challenge Dataset converted from JSON into
 R list structure}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

