% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDataFrame.R
\name{convertBusiness}
\alias{convertBusiness}
\title{The convertBusiness function
Convert from JSON file to data frame for business data}
\usage{
convertBusiness(data)
}
\arguments{
\item{data}{A JSON file of business information}
}
\description{
The convertBusiness function
Convert from JSON file to data frame for business data
}
\keyword{conversion}

