% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{convertDate}
\alias{convertDate}
\title{The convertDate function}
\usage{
convertDate(data)
}
\arguments{
\item{data}{vector of dates in character format year-month format}
}
\value{
returns a date vector
}
\description{
Converts from character year & month ex. ("2008-11") to date format
}
\author{
Ken Calhoon
}
\keyword{date}

