% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDataFrame.R
\name{convertReview}
\alias{convertReview}
\title{The convertReview function
Convert from JSON file to data frame for review data}
\usage{
convertReview(data)
}
\arguments{
\item{data}{A JSON file of review information}
}
\description{
The convertReview function
Convert from JSON file to data frame for review data
}
\keyword{conversion}

