% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDataFrame.R
\name{convertUser}
\alias{convertUser}
\title{The convertUser function
Convert from JSON file to data frame for user data}
\usage{
convertUser(data)
}
\arguments{
\item{data}{A JSON file of user information}
}
\description{
The convertUser function
Convert from JSON file to data frame for user data
}
\keyword{conversion}

