% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{featActiveDays}
\alias{featActiveDays}
\title{The featActiveDays function}
\usage{
featActiveDays(data, compDate = "2016-01-01")
}
\arguments{
\item{data}{vector of dates from which the number of days since compDate}

\item{compDate}{(optional) the date from which the input date is compared}
}
\description{
Returns an integer vector with number of days since compDate
}
\author{
Ken Calhoon
}
\keyword{feature}

