% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{featPredFemale}
\alias{featPredFemale}
\title{The featPredFemale function}
\usage{
featPredFemale(data, yearRange = c(1950, 1995))
}
\arguments{
\item{data}{character vector of first names}

\item{yearRange}{(optional) integer vector of start and stop date for name comparision}
}
\value{
Returns boolean vector where 1 == female
}
\description{
uses \code{\link[gender]{gender}} to predict gender based on first name
}
\examples{
featPredFemale(c("Grace","Ken"))
}
\author{
Ken Calhoon
}
\keyword{feature}

