% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{featSumBusDataByUser}
\alias{featSumBusDataByUser}
\title{The featSumBusDataByUser function}
\usage{
featSumBusDataByUser(fulldata)
}
\arguments{
\item{fulldata}{dataframe of all reviews, business and basic user data 
be grouped by user_id}
}
\description{
For each user_id, summarizes and scores the restaurant activity across many
 attributes
}
\author{
Ken Calhoon
}
\keyword{feature}

