% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndJoin.R
\name{filterAndJoin}
\alias{filterAndJoin}
\title{Filtering business features and joining datasets}
\usage{
filterAndJoin(review_data, user_data, business_data, state.name = NA,
  city.name = NA, type = NA)
}
\arguments{
\item{review_data}{a data frame of review information.}

\item{user_data}{a data frame of user information.}

\item{business_data}{a data frame of business information.}

\item{state.name}{optional, specify the state of business to be filtered.}

\item{city.name}{optional, specify the city of business to be filtered.}

\item{type}{optional, specify the type (by a key word) of business to be filtered.}
}
\description{
Given key words of business (state, city, type), the function return 
the joined data of review, business and user according to the filter conditions.
}
\examples{
data(reviewSampleDataFrame)
data(busSampleDataFrame)
data(userSampleDataFrame)
joined <- filterAndJoin(reviewSampleDataFrame, userSampleDataFrame, busSampleDataFrame,
city.name = "Dravosburg")
str(joined)

}
\author{
Xueying Liu
}
\keyword{filter}
\keyword{merge}

