% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphScoreDiff.R
\name{graphScoreDiff}
\alias{graphScoreDiff}
\title{The graphScoreDiff function}
\usage{
graphScoreDiff(fullclusterdata)
}
\arguments{
\item{fullclusterdata}{dataframe that includes business review information and cluster
assignments}
}
\value{
histogram plot of score differentials
}
\description{
Creates histogram of differences between the maximum score for a cluster
  group and the minimum score for a different cluster group. Wide differences show 
  the clusters scoring a business differently
}
\examples{
data("join100Sample")
data("user100Sample")
clusters <- 4
userFeatures <- appendUserData(fulldata = join100Sample, userdata = user100Sample)
dataCluster <- addCluster(join100Sample,appendKmeansCluster(userFeatures, k = clusters))
graphScoreDiff(dataCluster)
rm("join100Sample","user100Sample","clusters","userFeatures","dataCluster")
}
\author{
Ken Calhoon
}
\keyword{plot}

