% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initUserProfile.R
\name{initUserProfile}
\alias{initUserProfile}
\title{Initializing user profile}
\usage{
initUserProfile(joined.data)
}
\arguments{
\item{joined.data}{a data frame of joined information converted from "filterAndJoin".}
}
\description{
The function takes an argument of the joined data converted from the 
"filterAndJoin" function, and initializes a user profile from the joined data.
}
\examples{
data(join100Sample)
userProfile <- initUserProfile(join100Sample)
str(userProfile)


}
\author{
Xueying Liu
}
\keyword{profile}
\keyword{user}

