% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_100_sample.R
\docType{data}
\name{join100Sample}
\alias{join100Sample}
\title{Sample joined data frame of top 100 users from Yelp Challenge Dataset}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

