% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToDataFrame.R
\name{listToDataFrame}
\alias{listToDataFrame}
\title{Converting list to data frame}
\usage{
listToDataFrame(data, type = "")
}
\arguments{
\item{data}{a JSON file of business, user, or review data.}

\item{type}{specify the data source type, either "business", "user", or "review".}
}
\description{
The function converts from R list object file to data frame for business, 
user, or review data.
}
\examples{
data(busSampleListData)
class(busSampleListData)
businessDF <- listToDataFrame(busSampleListData, type = "business")
class(businessDF)
str(businessDF)

}
\author{
Xueying Liu
}
\keyword{conversion}

