% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansClustering.R
\name{prepKmeansData}
\alias{prepKmeansData}
\title{The prepKmeansData function}
\usage{
prepKmeansData(userdata)
}
\arguments{
\item{userdata}{dataframe of user data}
}
\value{
dataframe ready for kmeans
}
\description{
Formats, scales (as appropriate) and selects variables for kmeans
}
\author{
Ken Calhoon
}
\keyword{data}

