% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomByKeyWords.R
\name{recomByKeyWords}
\alias{recomByKeyWords}
\title{Recommendations based on given features}
\usage{
recomByKeyWords(data, state.name = NA, city.name = NA, types = NA,
  topNum = 10)
}
\arguments{
\item{data}{a data frame with business information.}

\item{state.name}{optional, specify the state of business to be recommended.}

\item{city.name}{optional, specify the city of business to be recommended.}

\item{types}{optional, a vector of characters to specify the types of business.}

\item{topNum}{specify the number of recommendations to be displayed. Default to be 10.}
}
\description{
Given some key words about the business location or types, the function
returns the top recommendations for the user-specified condition.
}
\examples{
data(busSampleDataFrame)
recomByKeyWords(busSampleDataFrame, types = "restaurant")

}
\author{
Xueying Liu
}
\keyword{recommendation}

