% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterRecommend.R
\name{recommendByCluster}
\alias{recommendByCluster}
\title{The recommendByCluster function}
\usage{
recommendByCluster(joinedData, userData, numberReco = 20,
  minClusterReview = 2, clusters = 4)
}
\arguments{
\item{joinedData}{dataset of reviews, business and user information}

\item{userData}{basic user dataset before feature and cluster addition}

\item{numberReco}{(optional) number of recommendations returned}

\item{minClusterReview}{(optional) minimum number of reviews, by cluster,
for a business to be included on the list}

\item{clusters}{(optional) number of user clusters}
}
\value{
list (of datatables) of recommended businesses in descending order by cluster
}
\description{
Produces the top recommendations for each cluster (value) in a list
}
\examples{
data("join100Sample")
data("user100Sample")
clusters <- 4
recos <- recommendByCluster(joinedData = join100Sample, userData = user100Sample, 
  numberReco = 20, minClusterReview = 2, clusters = 4)
sapply(seq(clusters), function (x) {print(head(recos[[x]], n = 3)[,c("cluster_grp","name")])})
rm("clusters","recos","join100Sample","user100Sample")
}
\author{
Ken Calhoon
}
\keyword{data}

