% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_sample_data_frame.R
\docType{data}
\name{reviewSampleDataFrame}
\alias{reviewSampleDataFrame}
\title{Sample review data from Yelp Challenge Dataset converted from R list structure
to R data frame structure}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

