% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_business_json.R
\docType{data}
\name{sample_business.json}
\alias{sample_business.json}
\title{Sample raw JSON business data from Yelp Challenge Dataset}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

