% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_100_sample.R
\docType{data}
\name{user100Sample}
\alias{user100Sample}
\title{Sample initial user profile data of top 100 users from Yelp Challenge Dataset}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

