% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_sample_list_data.R
\docType{data}
\name{userSampleListData}
\alias{userSampleListData}
\title{Sample user data from Yelp Challenge Dataset converted from JSON into
 R list structure}
\source{
\url{http://www.yelp.com/dataset_challenge/}
}
\description{
More description....
}
\keyword{data}

